refactor useYjsStore
